refactor scheme
