+++ controllers/contacts/add.js
@@ -1,21 +1,7 @@
-const Joi = require("joi");
-
 const contactsOperations = require("../../models/contacts");
-
-const contactsSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.string().required(),
-});
 
 const add = async (req, res, next) => {
   try {
-    const { error } = contactsSchema.validate(req.body);
-    if (error) {
-      error.status = 400;
-      throw error;
-    }
-
     const result = await contactsOperations.addContact(req.body);
     res.status(201).json({
       status: "success",
